hir_def: move visibility queries from hir_ty to hir_def
